+++ tests/migration/test_convert_to_new_inheritance_model.py
@@ -48,4 +48,4 @@ class ConvertToNewInheritanceModel(unittest.TestCase):
             collection.drop_index(index)
 
         # 5. Recreate indexes
-        Animal.objects._ensure_indexes()
+        Animal.ensure_indexes()