+++ tests/test_queryset.py
@@ -3078,7 +3078,6 @@ class QuerySetTest(unittest.TestCase):
         self.assertEqual([1, 2, 3], numbers)
         Number.drop_collection()
 
-
     def test_ensure_index(self):
         """Ensure that manual creation of indexes works.
         """
@@ -3086,7 +3085,7 @@ class QuerySetTest(unittest.TestCase):
             message = StringField()
             meta = {'allow_inheritance': True}
 
-        Comment.objects.ensure_index('message')
+        Comment.ensure_index('message')
 
         info = Comment.objects._collection.index_information()
         info = [(value['key'],